Create README.md
This program checks whether an applicant is eligible for a job with the eligibilities being that the applicant must be a US citizen and should either have a bachelor's degree or at least two years of experience. Using the Bool variable it has allowed me to have the variables "citizen" and "Batchlor" to identify as either "true" or "false," then from there using if-else statements to determine if the individual is eligible for the job position or not.